This is a pretty big update because it makes the game much more playable than it was previously. The changes are:
- Fixed a bug in the AI code which made it always choose itself as the
player to target.
- Moved the decideAction prototype to the Game.h file so that it can be
called in runGame.c.
- Added code to runGame.c so that the AI is called for player's 2, 3 and
4, effectively making the game work as a single player experience. This
was decided because printing out multiple player's hands to the screen
would be impractical for multiple human players.
- Created a runGame function "printHandDecision" which prints the hand
for player 1 only.
- Added an if statement to get the action from user input if the current
player is player one, or to print the AI's decisions otherwise, so
that the player knows what moves took place.
- Modified turnPrompt to add a manual delay using a for loop.

Some bugs that I noticed in testing which need to be fixed in the future:
- The checkPlayer or checkOpponent function is broken. It's possible that
it doesn't work when the player has multiple of the same card, since
that seems to trigger.
- The turnPrompt function still needs a proper solution to pause until
the user presses a key or enter.