This commit included formatting changes to properly display cards as A, J, Q and K if they are not numbered cards, and also to display the card that the AI asks for with the same notation. This involved using switch statements to convert the cards from integer to char and back.
One issue was that a 10 couldn't be converted to a character, so if the
AI asks for a 10, it is replaced with an X.

Since all of the original goals of the project have been met, this
release will be tagged as version 1.0. Any additions will be extra to
the original functionality.